Metrics
This commit contains two functions that return
   - The number of orders for a given event and its cancellation rate (as a percentage)
   - The date with the highest number of cancelled tickets (not orders! i.e: 1 order of 15 tickets > 14 orders of 1 ticket)

I am not sure what is the purpose of this part of the exercise. If it is to validate
whether the domain change allows to retrieve this information, then this solution
shoudl be fine.

Normally metrics would be taken out of a datawarehouse where the data is normalised
in order to quickly return responses. Here I am simply querying a "production" db
to show that the responses can be delivered with this domain solution.