Add attributes to models
Add attribute to state to Order and state_change_at to record
when the change of state happened.

Remove fulfulled as attribute of Order, fulfulled will
be one of the possible states.

All datetime fields are added as null=True to prevent
adding defaults for existing fields during the migration.
In following migrations the null value will be removed.